zcash_client_backend: Move data_api::testing module into its own file.
